Стой! Кто идёт?! : доработки
