+++ frontend/controllers/SecuredController.php
@@ -1,0 +1,25 @@
+<?php
+
+
+namespace frontend\controllers;
+
+use yii\filters\AccessControl;
+use yii\web\Controller;
+
+abstract class SecuredController extends Controller
+{
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::class,
+                'rules' => [
+                    [
+                        'allow' => true,
+                        'roles' => ['@']
+                    ]
+                ]
+            ]
+        ];
+    }
+}
