+++ frontend/controllers/TasksController.php
@@ -4,7 +4,6 @@
 namespace frontend\controllers;
 
 use frontend\models\tasks\Tasks;
-use yii\web\Controller;
 use yii\data\Pagination;
 use frontend\models\tasks\TasksFilterForm;
 use yii;
@@ -13,8 +12,9 @@ use yii\web\HttpException;
 /**
  * Tasks controller
  */
-class TasksController extends Controller
+class TasksController extends SecuredController
 {
+
     /**
      * Отображение общего списка заданий с учетом фильтров, если они заданы.
      *
