+++ frontend/models/tasks/Tasks.php
@@ -6,6 +6,7 @@ use Yii;
 use frontend\models\Categories;
 use frontend\models\Cities;
 use frontend\models\users\Users;
+use frontend\models\Files;
 
 /**
  * This is the model class for table "tasks".
