+++ frontend/models/users/Users.php
@@ -44,7 +44,7 @@ class Users extends \yii\db\ActiveRecord implements IdentityInterface
      */
     public static function tableName()
     {
-        return 'user';
+        return 'users';
     }
 
     /**
