+++ frontend/views/site/registration.php
