+++ frontend/views/site_old/index.php
@@ -1,0 +1,1 @@
+<?php
